% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array.R
\name{plot_array}
\alias{plot_array}
\title{plot array}
\usage{
plot_array(
  x = seq(iris$Sepal.Length),
  y = list(a = iris$Sepal.Length, b = iris$Sepal.Width, c = iris$Petal.Length),
  xim = range(x),
  ylim = range(y),
  xlab = "x",
  ylab = "y",
  subtitle = names(y)
)
}
\description{
plot array
}
